binários funcionam, tenho alguns warnings quando compila mas ta tudo "certo", estou com alguns problemas com relacaoa obter o cpf e o ano, se eu lidar com eles como se fossem arrays eu nao teria esse problema, mesmo assima cho que a raiz dele é a inicialização de variaveis?

+++ projeto_escola/interface_main.h
@@ -3,7 +3,7 @@
 
 struct data {
     unsigned long long int timestamp;
-    int ano;
+    unsigned long long int ano;
     int mes;
     int dia;
 };
@@ -13,7 +13,7 @@ struct aluno {
     char nome[256];
     char sexo;
     struct data nascimento;
-    int cpf;
+    unsigned long long int cpf;
 };
 
 struct professor {
@@ -21,7 +21,7 @@ struct professor {
     char nome[256];
     char sexo;
     struct data nascimento;
-    int cpf;
+    unsigned long long int cpf;
 };
 
 struct disciplina {
